Search for CSV in calling script's directory
When importing a module $PSScriptRoot is set to the module's directory not that of the calling script. Thus when you import remotely then look for $PSScriptRoot/machineConfig.CSV you aren't going to pull the config from the same directory as the calling script. This is IMHO a 'must have' feature for remotely.